pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gpt-thinking-extractor"
version = "0.1.0"
description = "Tools to scrape and extract thinking process data from ChatGPT threads"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.scripts]
gpt-scrape = "gpt_thinking_extractor.scrape_thoughts_final:run"
gpt-scrape-gui = "gpt_thinking_extractor.scraper_gui:main"

[tool.hatch.build.targets.wheel]
packages = ["src/gpt_thinking_extractor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]